Fix remaining low-severity audit findings and delete bugs.md
- Classify AddrNotAvailable as network error, not DNS failure (poller,
  eggpool); detect resolver failures via the getaddrinfo message prefix
- Await scheduler send on both fallback refresh paths so a full channel
  applies backpressure and closure surfaces through the error boundary
- Compare staleness age in u128 to remove the silent u64 truncation
- State the precise invariant in the filetime_to_u64 safety comment
- Hoist shared checks out of capable/incapable arms in V2 iowait, load,
  swap, and commit validators
- Add sample_interval_ms/observed_at_unix_ms setters to V2 builders and
  drop the drives clone in build_payload
- Document drive-name NUL exemption and truncate_width tiny-budget
  behavior
- Cover zero-total V2 drives, negative memory/swap/commit percentages,
  and negative iowait with regression tests

+++ crates/greggd/src/server/mod.rs
@@ -283,8 +283,7 @@ impl ServerState {
         if !self.max_snapshot_age.is_zero() {
             if let Some(observed_at_unix_ms) = state.last_observed_at_unix_ms {
                 let age_ms = now_unix_ms.saturating_sub(observed_at_unix_ms);
-                #[allow(clippy::cast_possible_truncation)]
-                if age_ms >= self.max_snapshot_age.as_millis() as u64 {
+                if u128::from(age_ms) >= self.max_snapshot_age.as_millis() {
                     return true;
                 }
             }
